add auth

+++ controllers/feed.js
@@ -4,17 +4,30 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
+  const currentPage = req.query.page || 1;
+  const perPage = 2;
+  let totalItems;
   Post
     .find()
+    .countDocuments()
+    .then(count => {
+      totalItems = count;
+      return Post
+        .find()
+        .skip((currentPage - 1) * perPage)
+        .limit(perPage);
+    })
     .then(posts => {
       res
         .status(200)
         .json({
           message: 'Fetched posts successfully!',
-          posts: posts
-        })
+          posts: posts,
+          totalItems: totalItems
+        });
     })
     .catch(err => {
       if (!err.statusCode) {
@@ -41,21 +54,34 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
   const imageUrl = req.file.path.replace('\\', '/');
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: 'Han Tran'
-    }
+    creator: req.userId
   });
   post
     .save()
     .then(result => {
-      res.status(201).json({
-        message: 'Post created successfully!',
-        post: result
-      });
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
+      res
+        .status(201)
+        .json({
+          message: 'Post created successfully!',
+          post: post,
+          creator: {
+            _id: creator._id,
+            name: creator.name
+          }
+        });
     })
     .catch(err => {
       if (!err.statusCode) {
@@ -121,6 +147,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -155,11 +186,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
-      Post.findByIdAndRemove(postId);
+      return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res
         .status(200)
         .json({
